Split up replay.ex
