+++ lib/battle_snake/replay.ex
@@ -3,6 +3,10 @@ defmodule BattleSnake.Replay do
   alias BattleSnake.Replay.Recorder
   alias BattleSnake.Replay.PlayBack
 
+  @moduledoc """
+  Top level API for Replays.
+  """
+
   defdelegate recorder_name(game_id), to: Replay.Registry
   defdelegate play_back_name(game_id), to: Replay.Registry
 
@@ -26,78 +30,5 @@ defmodule BattleSnake.Replay do
       {:error, {:already_started, pid}} -> {:ok, pid}
       {:ok, pid} -> {:ok, pid}
     end
-  end
-end
-
-defmodule BattleSnake.Replay.Registry do
-  def recorder_name(game_id) do
-    {:via, Registry, {BattleSnake.Replay.Registry, "recorder:#{game_id}"}}
-  end
-
-  def play_back_name(game_id) do
-    {:via, Registry, {BattleSnake.Replay.Registry, "playback:#{game_id}"}}
-  end
-end
-
-defmodule BattleSnake.Replay.PlayBack.Supervisor do
-  use Supervisor
-  alias BattleSnake.Replay
-
-  def start_link do
-    Supervisor.start_link(__MODULE__, :ok, name: __MODULE__)
-  end
-
-  def start_child(opts) do
-    Supervisor.start_child(__MODULE__, opts)
-  end
-
-  def init(:ok) do
-    children = [
-      worker(Replay.PlayBack, [], restart: :temporary)
-    ]
-
-    supervise(children, strategy: :simple_one_for_one)
-  end
-end
-
-defmodule BattleSnake.Replay.Recorder.Supervisor do
-  use Supervisor
-  alias BattleSnake.Replay
-
-  def start_link do
-    Supervisor.start_link(__MODULE__, :ok, name: __MODULE__)
-  end
-
-  def start_child(opts) do
-    Supervisor.start_child(__MODULE__, opts)
-  end
-
-  def init(:ok) do
-    children = [
-      worker(Replay.Recorder, [], restart: :temporary)
-    ]
-
-    supervise(children, strategy: :simple_one_for_one)
-  end
-end
-
-defmodule BattleSnake.Replay.Supervisor do
-  use Supervisor
-  alias BattleSnake.Replay
-
-  def start_link do
-    Supervisor.start_link(__MODULE__, :ok, name: __MODULE__)
-  end
-
-  def init(:ok) do
-    children = [
-      supervisor(Replay.PlayBack.Supervisor, []),
-      supervisor(Replay.Recorder.Supervisor, []),
-      supervisor(Registry,
-        [:unique, BattleSnake.Replay.Registry],
-        [id: BattleSnake.Replay.Registry])
-    ]
-
-    supervise(children, strategy: :one_for_one)
   end
 end
